+++ coordinator/src/services/translation/request/dispense/dispense-notification.ts
@@ -1,12 +1,10 @@
 import {fhir, hl7V3} from "@models"
 import {
-  getCodingForSystem,
   getExtensionForUrl,
   getIdentifierValueForSystem,
   getIdentifierValueOrNullForSystem,
   getMessageId,
   getMedicationCoding,
-  onlyElement,
   getExtensionForUrlOrNull,
   getNumericValueAsString
 } from "../../common"
@@ -15,8 +13,17 @@ import {convertIsoDateTimeStringToHl7V3DateTime, convertMomentToHl7V3DateTime} f
 import pino from "pino"
 import {createAgentPersonForUnattendedAccess} from "../agent-unattended"
 import moment from "moment"
-import {auditDoseToTextIfEnabled} from "../dosage"
-import {dispenseNotificationIsRepeatDispensing} from "../index"
+import {dispenseIsRepeatDispensing} from "../index"
+import {DispenseNotificationPertinentSupplyHeader} from "../../../../../../models/hl7-v3"
+import {
+  createAgentOrganisation,
+  createPertinentInformation1LineItem,
+  createPertinentPrescriptionId,
+  createPertinentPrescriptionStatus,
+  createPriorOriginalRef,
+  createPriorPrescriptionReleaseEventRef,
+  getOrganisationPerformer
+} from "./dispense-common"
 
 export async function convertDispenseNotification(
   bundle: fhir.Bundle,
@@ -51,7 +58,7 @@ export async function convertDispenseNotification(
   hl7DispenseNotification.effectiveTime = convertMomentToHl7V3DateTime(moment.utc())
   hl7DispenseNotification.recordTarget = new hl7V3.RecordTargetReference(hl7Patient)
   hl7DispenseNotification.primaryInformationRecipient =
-    new hl7V3.DispenseNotificationPrimaryInformationRecipient(hl7AgentOrganisation)
+    new hl7V3.DispensePrimaryInformationRecipient(hl7AgentOrganisation)
   hl7DispenseNotification.pertinentInformation1 = hl7PertinentInformation1
   hl7DispenseNotification.pertinentInformation2 = new hl7V3.DispensePertinentInformation2(hl7CareRecordElementCategory)
   if (hl7PriorMessageRef) {
@@ -89,14 +96,13 @@ async function createPertinentInformation1(
       )
     }
   )
-  const supplyHeader = new hl7V3.PertinentSupplyHeader(new hl7V3.GlobalIdentifier(messageId))
-  supplyHeader.author = hl7Author
+  const supplyHeader = new DispenseNotificationPertinentSupplyHeader(new hl7V3.GlobalIdentifier(messageId), hl7Author)
   supplyHeader.pertinentInformation1 = hl7PertinentInformation1LineItems
   supplyHeader.pertinentInformation3 = new hl7V3.DispensePertinentInformation3(hl7PertinentPrescriptionStatus)
   supplyHeader.pertinentInformation4 = new hl7V3.DispensePertinentInformation4(hl7PertinentPrescriptionIdentifier)
   supplyHeader.inFulfillmentOf = new hl7V3.InFulfillmentOf(hl7PriorOriginalRef)
 
-  if (dispenseNotificationIsRepeatDispensing(fhirFirstMedicationDispense)) {
+  if (dispenseIsRepeatDispensing(fhirFirstMedicationDispense)) {
     const repeatInfo = getExtensionForUrlOrNull(
       fhirFirstMedicationDispense.extension,
       "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation",
@@ -116,111 +122,7 @@ async function createPertinentInformation1(
     )
   }
 
-  return new hl7V3.DispenseNotificationPertinentInformation1(supplyHeader)
-}
-
-function createPertinentInformation1LineItem(
-  fhirMedicationDispense: fhir.MedicationDispense,
-  medicationCoding: fhir.Coding,
-  logger: pino.Logger
-) {
-  const fhirPrescriptionDispenseItemNumber = getPrescriptionItemNumber(fhirMedicationDispense)
-  const fhirPrescriptionLineItemStatus = getPrescriptionLineItemStatus(fhirMedicationDispense)
-  const fhirDosageInstruction = getDosageInstruction(fhirMedicationDispense, logger)
-
-  const hl7SuppliedLineItemQuantitySnomedCode = new hl7V3.SnomedCode(
-    fhirMedicationDispense.quantity.code,
-    fhirMedicationDispense.quantity.unit
-  )
-  const hl7UnitValue = fhirMedicationDispense.quantity.value.toString()
-  const hl7Quantity = new hl7V3.QuantityInAlternativeUnits(
-    hl7UnitValue,
-    hl7UnitValue,
-    hl7SuppliedLineItemQuantitySnomedCode
-  )
-  const hl7ItemStatusCode = createLineItemStatusCode(fhirPrescriptionLineItemStatus)
-  const hl7PriorOriginalItemRef = getPrescriptionItemId(fhirMedicationDispense)
-  const hl7SuppliedLineItemQuantity = createSuppliedLineItemQuantity(
-    hl7SuppliedLineItemQuantitySnomedCode,
-    hl7Quantity,
-    medicationCoding,
-    fhirDosageInstruction
-  )
-
-  const hl7PertinentSuppliedLineItem = new hl7V3.PertinentSuppliedLineItem(
-    new hl7V3.GlobalIdentifier(fhirPrescriptionDispenseItemNumber),
-    new hl7V3.SnomedCode(medicationCoding.code)
-  )
-  hl7PertinentSuppliedLineItem.consumable = new hl7V3.Consumable(
-    new hl7V3.RequestedManufacturedProduct(
-      new hl7V3.ManufacturedRequestedMaterial(
-        hl7SuppliedLineItemQuantitySnomedCode
-      )
-    )
-  )
-  hl7PertinentSuppliedLineItem.component = new hl7V3.DispenseLineItemComponent(hl7SuppliedLineItemQuantity)
-  hl7PertinentSuppliedLineItem.component1 = new hl7V3.DispenseLineItemComponent1(
-    new hl7V3.SupplyRequest(hl7SuppliedLineItemQuantitySnomedCode, hl7Quantity)
-  )
-  hl7PertinentSuppliedLineItem.pertinentInformation3 = new hl7V3.DispenseLineItemPertinentInformation3(
-    new hl7V3.PertinentItemStatus(hl7ItemStatusCode)
-  )
-  hl7PertinentSuppliedLineItem.inFulfillmentOf = new hl7V3.InFulfillmentOfLineItem(
-    new hl7V3.PriorOriginalRef(new hl7V3.GlobalIdentifier(hl7PriorOriginalItemRef))
-  )
-
-  if (dispenseNotificationIsRepeatDispensing(fhirMedicationDispense)) {
-    const repeatInfo = getExtensionForUrlOrNull(
-      fhirMedicationDispense.extension,
-      "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation",
-      "MedicationDispense.extension"
-    )
-    const numberOfRepeatsAllowedExtension = getExtensionForUrl(
-      repeatInfo.extension,
-      "numberOfRepeatsAllowed",
-      /* eslint-disable-next-line max-len */
-      'MedicationDispense.extension("https://fhir.nhs.uk/StructureDefinition/Extension-EPS-RepeatInformation").extension'
-    ) as fhir.IntegerExtension
-    const numberOfRepeatsAllowed = getNumericValueAsString(numberOfRepeatsAllowedExtension.valueInteger)
-
-    hl7PertinentSuppliedLineItem.repeatNumber = new hl7V3.Interval<hl7V3.NumericValue>(
-      new hl7V3.NumericValue(undefined),
-      new hl7V3.NumericValue(numberOfRepeatsAllowed)
-    )
-  }
-
-  return new hl7V3.DispenseNotificationPertinentInformation1LineItem(hl7PertinentSuppliedLineItem)
-}
-
-function createSuppliedLineItemQuantity(
-  hl7SuppliedLineItemQuantitySnomedCode: hl7V3.SnomedCode,
-  hl7Quantity: hl7V3.QuantityInAlternativeUnits,
-  fhirMedicationCodeableConceptCoding: fhir.Coding,
-  fhirDosageInstruction: fhir.Dosage
-) {
-  const hl7SuppliedLineItemQuantity = new hl7V3.SuppliedLineItemQuantity()
-  hl7SuppliedLineItemQuantity.code = hl7SuppliedLineItemQuantitySnomedCode
-  hl7SuppliedLineItemQuantity.quantity = hl7Quantity
-  hl7SuppliedLineItemQuantity.product = new hl7V3.DispenseProduct(
-    new hl7V3.SuppliedManufacturedProduct(
-      new hl7V3.ManufacturedRequestedMaterial(
-        new hl7V3.SnomedCode(
-          fhirMedicationCodeableConceptCoding.code,
-          fhirMedicationCodeableConceptCoding.display
-        )
-      )
-    )
-  )
-  hl7SuppliedLineItemQuantity.pertinentInformation1 = new hl7V3.DispenseLineItemPertinentInformation1(
-    new hl7V3.PertinentSupplyInstructions(
-      new hl7V3.Text(fhirDosageInstruction.text)
-    )
-  )
-  return hl7SuppliedLineItemQuantity
-}
-
-export function getOrganisationPerformer(fhirFirstMedicationDispense: fhir.MedicationDispense): fhir.DispensePerformer {
-  return fhirFirstMedicationDispense.performer.find(p => p.actor.type === "Organization")
+  return new hl7V3.DispensePertinentInformation1(supplyHeader)
 }
 
 function getLineItemIdentifiers(fhirMedicationDispenses: Array<fhir.MedicationDispense>) {
@@ -242,120 +144,6 @@ function getNhsNumber(fhirPatient: fhir.Patient, fhirFirstMedicationDispense: fh
     : fhirFirstMedicationDispense.subject.identifier.value
 }
 
-export function getFhirGroupIdentifierExtension(
-  fhirFirstMedicationDispense: fhir.MedicationDispense
-): fhir.ExtensionExtension<fhir.Extension> {
-  const fhirAuthorizingPrescriptionExtensions =
-    fhirFirstMedicationDispense.authorizingPrescription.flatMap(e => e.extension)
-  const fhirGroupIdentifierExtension = getExtensionForUrl(
-    fhirAuthorizingPrescriptionExtensions,
-    "https://fhir.nhs.uk/StructureDefinition/Extension-DM-GroupIdentifier",
-    "MedicationDispense.authorizingPrescription") as fhir.ExtensionExtension<fhir.Extension>
-  return fhirGroupIdentifierExtension
-}
-
-export function getPrescriptionStatus(fhirFirstMedicationDispense: fhir.MedicationDispense): fhir.CodingExtension {
-  return getExtensionForUrl(
-    fhirFirstMedicationDispense.extension,
-    "https://fhir.nhs.uk/StructureDefinition/Extension-EPS-TaskBusinessStatus",
-    "MedicationDispense.extension") as fhir.CodingExtension
-}
-
-function getPrescriptionItemId(fhirMedicationDispense: fhir.MedicationDispense) {
-  return getIdentifierValueForSystem(
-    fhirMedicationDispense.authorizingPrescription.map(e => e.identifier),
-    "https://fhir.nhs.uk/Id/prescription-order-item-number",
-    "MedicationDispense.authorizingPrescription.identifier"
-  )
-}
-
-function getDosageInstruction(fhirMedicationDispense: fhir.MedicationDispense, logger: pino.Logger) {
-  auditDoseToTextIfEnabled(fhirMedicationDispense.dosageInstruction, logger)
-  return onlyElement(
-    fhirMedicationDispense.dosageInstruction,
-    "MedicationDispense.dosageInstruction"
-  )
-}
-
-export function getPrescriptionItemNumber(fhirMedicationDispense: fhir.MedicationDispense): string {
-  return getIdentifierValueForSystem(
-    fhirMedicationDispense.identifier,
-    "https://fhir.nhs.uk/Id/prescription-dispense-item-number",
-    "MedicationDispense.identifier"
-  )
-}
-
-function getPrescriptionLineItemStatus(fhirMedicationDispense: fhir.MedicationDispense) {
-  return getCodingForSystem(
-    fhirMedicationDispense.type.coding,
-    "https://fhir.nhs.uk/CodeSystem/medicationdispense-type",
-    "MedicationDispense.type.coding"
-  )
-}
-
-export function createLineItemStatusCode(fhirPrescriptionLineItemStatus: fhir.Coding): hl7V3.ItemStatusCode {
-  const itemStatusCode = new hl7V3.ItemStatusCode(fhirPrescriptionLineItemStatus.code)
-  itemStatusCode._attributes.displayName = fhirPrescriptionLineItemStatus.display
-  return itemStatusCode
-}
-
-function createPertinentPrescriptionId(fhirFirstMedicationDispense: fhir.MedicationDispense) {
-  const fhirGroupIdentifierExtension = getFhirGroupIdentifierExtension(fhirFirstMedicationDispense)
-  const fhirAuthorizingPrescriptionShortFormIdExtension = getExtensionForUrl(
-    fhirGroupIdentifierExtension.extension,
-    "shortForm",
-    "MedicationDispense.authorizingPrescription.extension.valueIdentifier"
-  ) as fhir.IdentifierExtension
-  const hl7PertinentPrescriptionId = fhirAuthorizingPrescriptionShortFormIdExtension
-    .valueIdentifier
-    .value
-  return new hl7V3.PrescriptionId(hl7PertinentPrescriptionId)
-}
-
-function createPriorOriginalRef(firstMedicationDispense: fhir.MedicationDispense) {
-  const fhirGroupIdentifierExtension = getFhirGroupIdentifierExtension(firstMedicationDispense)
-  const fhirAuthorizingPrescriptionShortFormIdExtension = getExtensionForUrl(
-    fhirGroupIdentifierExtension.extension,
-    "UUID",
-    "MedicationDispense.authorizingPrescription.extension.valueIdentifier"
-  ) as fhir.IdentifierExtension
-  const id = fhirAuthorizingPrescriptionShortFormIdExtension.valueIdentifier.value
-  return new hl7V3.PriorOriginalRef(
-    new hl7V3.GlobalIdentifier(id)
-  )
-}
-
-export function createPertinentPrescriptionStatus(
-  fhirFirstMedicationDispense: fhir.MedicationDispense
-): hl7V3.PertinentPrescriptionStatus {
-  const fhirPrescriptionStatus = getPrescriptionStatus(fhirFirstMedicationDispense)
-  const hl7StatusCode = new hl7V3.StatusCode(fhirPrescriptionStatus.valueCoding.code)
-  hl7StatusCode._attributes.displayName = fhirPrescriptionStatus.valueCoding.display
-  return new hl7V3.PertinentPrescriptionStatus(hl7StatusCode)
-}
-
-function createPatient(patient: fhir.Patient, firstMedicationDispense: fhir.MedicationDispense): hl7V3.Patient {
-  const nhsNumber = getNhsNumber(patient, firstMedicationDispense)
-  const hl7Patient = new hl7V3.Patient()
-  hl7Patient.id = new hl7V3.NhsNumber(nhsNumber)
-  return hl7Patient
-}
-
-function createAgentOrganisation(organisation: fhir.DispensePerformer): hl7V3.AgentOrganization {
-  const organisationCode = organisation.actor.identifier.value
-  const organisationName = organisation.actor.display
-  const hl7Organisation = createOrganisation(organisationCode, organisationName)
-  return new hl7V3.AgentOrganization(hl7Organisation)
-}
-
-function createOrganisation(organisationCode: string, organisationName: string): hl7V3.Organization {
-  const organisation = new hl7V3.Organization()
-  organisation.id = new hl7V3.SdsOrganizationIdentifier(organisationCode)
-  organisation.code = new hl7V3.OrganizationTypeCode()
-  organisation.name = new hl7V3.Text(organisationName)
-  return organisation
-}
-
 async function createAuthor(
   organisationCode: string,
   authorTime: string,
@@ -366,6 +154,13 @@ async function createAuthor(
   author.signatureText = hl7V3.Null.NOT_APPLICABLE
   author.AgentPerson = await createAgentPersonForUnattendedAccess(organisationCode, logger)
   return author
+}
+
+function createPatient(patient: fhir.Patient, firstMedicationDispense: fhir.MedicationDispense): hl7V3.Patient {
+  const nhsNumber = getNhsNumber(patient, firstMedicationDispense)
+  const hl7Patient = new hl7V3.Patient()
+  hl7Patient.id = new hl7V3.NhsNumber(nhsNumber)
+  return hl7Patient
 }
 
 function createCareRecordElementCategory(fhirIdentifiers: Array<string>) {
@@ -394,10 +189,4 @@ function createPriorMessageRef(fhirHeader: fhir.MessageHeader) {
   if (replacementOf) {
     return new hl7V3.MessageRef(new hl7V3.GlobalIdentifier(replacementOf.valueIdentifier.value))
   }
-}
-
-function createPriorPrescriptionReleaseEventRef(fhirHeader: fhir.MessageHeader) {
-  return new hl7V3.PriorPrescriptionReleaseEventRef(
-    new hl7V3.GlobalIdentifier(fhirHeader.response.identifier)
-  )
 }
