+++ models/hl7-v3/index.ts
@@ -4,6 +4,7 @@ export * from "./cancellation-response"
 export * from "./codes"
 export * from "./core"
 export * from "./demographics"
+export * from "./dispense-common"
 export * from "./dispense-notification"
 export * from "./dispense-claim-information"
 export * from "./line-item"
